auth_jwt/api/serializers/TransactionSerializer.py
from rest_framework import serializers
from api.models.Transaction import Transaction

class TransactionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Transaction
        fields = '__all__'