auth_jwt/requirements.txt
asgiref==3.8.1
distlib==0.3.9
Django==5.0.3
django-allauth==65.0.2
django-filter==24.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==30.8.1
filelock==3.16.1
Markdown==3.7
platformdirs==4.3.6
PyJWT==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==0.20.0
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.12.2
tzdata==2024.2
validate-docbr==1.10.0
virtualenv==20.26.6
